partial move to snapshot-service crate
I kept the filesystem watcher in vit-servicing-station-lib because I
don't think it's a core part of the API, and for something like mocking
should the API be enough. Also, because I'm considering exposing here an
extra endpoint to do the update protected with an api key.

I also removed the integration with the DB because it was easier to do
it this way first. I'm considering using sled instead of sqlite also,
but the in-memory data structures are probably enough anyway for our
use-cases I think.