Voting groups api (#229)
# API

## Breaking changes

### proposals -> proposal

`GET v0/proposals/{id}` is now `GET v0/proposal/{id}/{group-id}`

I removed the `s` since it returns a single proposal, which I think is more consistent. Also added the group-id filter.

### get all proposals

`GET v0/proposals` is now `GET v0/proposals/{group-id}`

## Comments

`POST v0/proposals` is kept as it is.

# Data

Add new two tables

# proposals_voteplans

As was more of less hinted in #200 I added a 1-N relationship between ideascale proposals and on-chain proposals. I kept a unified view in the API though, because it's simpler.

This also adds a new csv to load this table. I think this approach is cleaner than duplicating each proposal but changing voteplan ids. The only cons I see with this is that as it is right now an ideascale proposal could be not associated with any on chain proposal if we mess up the construction, but we could add a check for that anyway.

# groups

Similar thing, this table maps tokens to groups names/ids. Also has its own new csv file to populate it. 

# Use-cases/flow

As it is right now, this is how I understand things should work.

Users can know which groups they belong to in two ways: 

One is that they fetch the account status from the node to get their voting tokens. Then fetch the fund with `v0/fund`, and filter by this (on client side)

Two is that they use the snapshot endpoint once it's finished, since it will provide the voting power for each group, along with the group id.

Then they can operate within one group at a time, by adding the parameter to the `proposal` endpoint. The rest should be equal to the way things work right now.

Co-authored-by: Yevhenii Babichenko <eugene.babichenko@iohk.io>